Various improvements to the debug console
The debug console will first attempt to eval the statement, then if that fails it will exec it. This allows creating variables and functions in the debug console, as well as assigning values to various parts of the game.

Event propagation on the <Return> event for the input area of the debug is now disabled.

Scroll bars have been added to both the input and output areas of the debug console.