Changed Config type
The old config type had a lot of hardcoded assumptions, and was not easy
to use. It also did all checking of values at access time. This updates
the config type to be defined once, initially, and have valid values at
all times. The interface for grabbing values was also changed to exploit
the __getattr__ magic method, allowing for nice python-like syntax for
the dynamic values.