Code-style cleanup (flake8 and mypy)
This cleans up the code such that a run of flake8 on the ./rpg directory does
not produce any errors or warnings. This also takes care of several errors that
mypy identifies.

This was done on windows, but with the git option to ignore the executable bit
set - hopefully this means that git won't clobber the file mode for all the
changed files.