Split items.py into submodule, update Attributes to use properties
This also updates the secondary attributes (formerly called Stats) to
automatically track changes to the primary attributes that govern their values.
Some tests for this behavior were added as well which is something I'm going to
be doing more of, as this helped me nail down the behavior that was expected in
a way that should reduce possible bugs.